requirements.txt
asgiref==3.5.2
async-generator==1.10
attrs==21.4.0
black==22.3.0
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
cryptography==36.0.1
distlib==0.3.4
dj-database-url==0.5.0
Django==4.0.5
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
django-on-heroku==1.1.2
djangorestframework==3.13.1
filelock==3.7.0
gunicorn==20.1.0
h11==0.13.0
idna==3.3
iniconfig==1.1.1
mypy-extensions==0.4.3
outcome==1.1.0
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.2
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pycparser==2.21
pyOpenSSL==22.0.0
pyparsing==3.0.8
pytest==7.1.2
pytest-django==4.5.2
python-decouple==3.6
pytz==2022.1
requests==2.28.0
selenium==4.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
tomli==2.0.1
trio==0.19.0
trio-websocket==0.9.2
typing_extensions==4.2.0
tzdata==2021.5
urllib3==1.26.9
virtualenv==20.14.1
whitenoise==6.2.0
wsproto==1.0.0
